requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
PyPDF2
python-docx
tqdm